Fix a problem with removing items from the mini cart
Every time the cart is updated, the mini cart HTML is regenerated and injected.
Since the mini-cart tracking JS is directly added to the HTML for the mini cart,
the $(document.body).on('click') event gets added again with every regeneration.
So each time you remove an item, another listener is added to the rest of the remove
buttons.