Maybe display GA Pro notice on new installation and when orders get completed
Previously, the shop's orders were counted on all page loads because it was hooked into `admin_init`. This was causing performance issues for some users even though WooCommerce caches the results. Given that the notice isn't critical, we should only check if it should be displayed whenever a new order gets completed (since we're counting the completed orders) or on plugin activation.